stock-manager/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
aiosqlite==0.19.0
python-multipart==0.0.6
httpx==0.25.2
python-telegram-bot==20.7